add input conditions

+++ src/controller/car-controller.js
@@ -7,6 +7,21 @@ const createslot = async (req, res) => {
         if (!req.body.parkingLotId || !req.body.registrationNumber || !req.body.color) {
             throw "parking Lot Id or ,reg no. or color missing";
         }
+        const parkingLotId = req.body.parkingLotId;
+        const registrationNumber = req.body.registrationNumber;
+        const color = req.body.color.toUpperCase();
+        const possiblecolor = ['RED', 'GREEN', 'BLUE', 'BLACK', 'WHITE', 'YELLOW', 'ORANGE'];
+        if (registrationNumber.length !== 9) {
+            throw 'Registration Number is not valid';
+        }
+        const distict = registrationNumber[2] + registrationNumber[3];
+        if (distict > "20") {
+            throw 'Registration Number is not valid';
+        }
+        if (!possiblecolor.includes(color)) {
+            throw 'Color is not valid';
+        }
+
         const response = await carService.create({
             "parkingLotId": req.body.parkingLotId,
             "registrationNumber": req.body.registrationNumber,
@@ -36,11 +51,22 @@ const leavecar = async (req, res) => {
         if (!req.body.parkingLotId || !req.body.registrationNumber) {
             throw "Parking Lot Id or registaration number missing";
         }
+        const parkingLotId = req.body.parkingLotId;
+        const registrationNumber = req.body.registrationNumber;
+        if (registrationNumber.length !== 9) {
+            throw 'Registration Number is not valid';
+        }
+        const distict = registrationNumber[2] + registrationNumber[3];
+        if (distict > "20") {
+            throw 'Registration Number is not valid';
+        }
+
+
         const response = await carService.remove({
             "parkingLotId": req.body.parkingLotId,
             "registrationNumber": req.body.registrationNumber
         });
-        console.log(response);
+
         return res.status(200).json({
             "isSuccess": true,
             "response": {
@@ -68,7 +94,7 @@ const getregistrations = async (req, res) => {
             throw 'Query parameter is missing';
         }
         const response = await carService.conditionalselect(
-            { color: color, parkingLotId: parkingLotId }, 'registrationNumber -_id');
+            { color: color, parkingLotId: parkingLotId }, 'registrationNumber -_id', '+updatedAt');
         return res.status(200).json({
             "isSuccess": true,
             "response": {
@@ -92,7 +118,7 @@ const getslots = async (req, res) => {
             throw 'Query parameter is missing';
         }
         const response = await carService.conditionalselect(
-            { color: color, parkingLotId: parkingLotId }, 'slotNumber -_id');
+            { color: color, parkingLotId: parkingLotId }, 'slotNumber -_id', '+slotNumber');
         return res.status(200).json({
             "isSuccess": true,
             "response": {
